concat

+++ tests/test_spatial_model.py
@@ -4,10 +4,10 @@ from src.models import spatial
 from pytorch_lightning import Trainer
 
 def test_fit():
-    train_sensor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10))
-    train_neighbor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10))
-    val_sensor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10))
-    val_neighbor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10))
+    train_sensor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10)).astype(np.float32)
+    train_neighbor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10)).astype(np.float32)
+    val_sensor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10)).astype(np.float32)
+    val_neighbor_score = np.random.multinomial(1,[1/6,2/6,3/6], (10)).astype(np.float32)
     train_labels = np.random.randint(low=0, high=3, size = 10)
     val_labels = np.random.randint(low=0, high=3, size = 10)
